Refactor Blog fetch
