+++ src/views/Presentation/components/Blogs/Blogs.js
@@ -1,97 +1,12 @@
-// import React from 'react';
-// import PropTypes from 'prop-types';
-// import clsx from 'clsx';
-// import { makeStyles } from '@material-ui/core/styles';
-// import { Typography, Grid } from '@material-ui/core';
-
-// const useStyles = makeStyles(theme => ({
-//   root: {
-//     backgroundColor: theme.palette.background.default
-//   },
-//   inner: {
-//     padding: theme.spacing(6, 3),
-//     width: theme.breakpoints.values.lg,
-//     maxWidth: '100%',
-//     margin: '0 auto'
-//   },
-//   media: {
-//     '& img': {
-//       width: '100%',
-//       height: 'auto'
-//     }
-//   }
-// }));
-
-// const UserFlows = props => {
-//   const { className, ...rest } = props;
-
-//   const classes = useStyles();
-
-//   return (
-//     <div
-//       {...rest}
-//       className={clsx(classes.root, className)}
-//     >
-//       <div className={classes.inner}>
-//         <Grid
-//           alignItems="center"
-//           container
-//           justify="space-between"
-//         >
-//           <Grid
-//             item
-//             lg={3}
-//             md={6}
-//             xs={12}
-//           >
-//             <Typography
-//               gutterBottom
-//               variant="h3"
-//             >
-//               Complete User Flows
-//             </Typography>
-//             <Typography variant="subtitle2">
-//               Not just a set of tools, the package includes the most common use
-//               cases of user flows like User Management, Second Level Layout, and
-//               many more.
-//             </Typography>
-//           </Grid>
-//           <Grid
-//             className={classes.media}
-//             item
-//             lg={4}
-//             md={6}
-//             xs={12}
-//           >
-//             <img
-//               alt="User flows"
-//               src="/images/presentation/user_flows.png"
-//             />
-//           </Grid>
-//         </Grid>
-//       </div>
-//     </div>
-//   );
-// };
-
-// UserFlows.propTypes = {
-//   className: PropTypes.string
-// };
-
-// export default UserFlows;
-
 import React from 'react';
 import cx from 'clsx';
 import Color from 'color';
 import GoogleFontLoader from 'react-google-font-loader';
 import NoSsr from '@material-ui/core/NoSsr';
 import { makeStyles } from '@material-ui/core/styles';
-import Avatar from '@material-ui/core/Avatar';
-import Box from '@material-ui/core/Box';
-import CardMedia from '@material-ui/core/CardMedia';
 import Grid from '@material-ui/core/Grid';
-import { useCoverCardMediaStyles } from '@mui-treasury/styles/cardMedia/cover';
-import { Row, Item } from '@mui-treasury/components/flex';
+import CustomCard from './CustomCard';
+import { Blogs } from '..';
 
 const useGridStyles = makeStyles(({ breakpoints }) => ({
   root: {
@@ -107,126 +22,33 @@ const useGridStyles = makeStyles(({ breakpoints }) => ({
   },
 }));
 
-const useStyles = makeStyles(({ palette }) => ({
-  color: ({ color }: { color: string }) => ({
-    '&:before': {
-      backgroundColor: Color(color)
-        .darken(0.3)
-        .desaturate(0.2)
-        .toString(),
-    },
-  }),
-  root: {
-    position: 'relative',
-    borderRadius: '1rem',
-    minWidth: 320,
-    '&:before': {
-      transition: '0.2s',
-      position: 'absolute',
-      width: '100%',
-      height: '100%',
-      content: '""',
-      display: 'block',
-      borderRadius: '1rem',
-      zIndex: 0,
-      bottom: 0,
-    },
-    '&:hover': {
-      '&:before': {
-        bottom: -6,
-      },
-      '& $logo': {
-        transform: 'scale(1.1)',
-        boxShadow: '0 6px 20px 0 rgba(0,0,0,0.38)',
-      },
-    },
+const blogs = [
+  {
+    usrImage: 'https://d3fa68hw0m2vcc.cloudfront.net/bf4/156511609.jpeg',
+    cover: 'https://cdn.vox-cdn.com/thumbor/C6_-SDnnoFdS19XRH4XvAYN1BT8=/148x0:1768x1080/1400x1400/filters:focal(148x0:1768x1080):format(jpeg)/cdn.vox-cdn.com/uploads/chorus_image/image/49641465/tracer_overwatch.0.0.jpg',
+    title: 'Astronomy Binoculars A Great Alternative',
+    description: 'Overwatch Official',
+    date: '02.04.2020',
+    color: '#fc7944'
   },
-  cover: {
-    borderRadius: '1rem',
+  {
+    usrImage: 'https://vignette.wikia.nocookie.net/youtube/images/7/77/LeagueOfLegends.jpg/revision/latest?cb=20180718040905',
+    cover: 'https://www.pcclean.io/wp-content/uploads/2019/04/559308.jpg',
+    title: 'New blogs for young people',
+    description: 'League of Legends Official',
+    date: '02.04.2020',
+    color: '#5357ce'
   },
-  content: ({ color }: { color: string }) => ({
-    position: 'relative',
-    zIndex: 1,
-    borderRadius: '1rem',
-    boxShadow: `0 6px 16px 0 ${Color(color).fade(0.5)}`,
-    '&:before': {
-      content: '""',
-      display: 'block',
-      position: 'absolute',
-      left: 0,
-      top: 0,
-      zIndex: 0,
-      width: '100%',
-      height: '100%',
-      clipPath:
-        'polygon(0% 100%, 0% 35%, 0.3% 33%, 1% 31%, 1.5% 30%, 2% 29%, 2.5% 28.4%, 3% 27.9%, 3.3% 27.6%, 5% 27%,95% 0%,100% 0%, 100% 100%)',
-      borderRadius: '1rem',
-      background: `linear-gradient(to top, ${color}, ${Color(color)
-        .rotate(24)
-        .lighten(0.12)})`,
-    },
-  }),
-  title: {
-    fontFamily: 'Fjalla One',
-    fontSize: '1.25rem',
-    color: '#fff',
-    margin: 0,
+  {
+    usrImage: 'https://vignette.wikia.nocookie.net/youtube/images/7/77/LeagueOfLegends.jpg/revision/latest?cb=20180718040905',
+    cover: 'https://www.pcclean.io/wp-content/uploads/2019/04/559308.jpg',
+    title: 'New blogs for young people',
+    description: 'League of Legends Official',
+    date: '02.04.2020',
+    color: '#7f5500'
   },
-  logo: {
-    transition: '0.3s',
-    width: 100,
-    height: 100,
-    boxShadow: '0 4px 12px 0 rgba(0,0,0,0.24)',
-    borderRadius: '1rem',
-  },
-  team: {
-    fontFamily: 'Sen',
-    fontSize: '0.75rem',
-    color: palette.text.hint,
-  },
-  date: {
-    fontFamily: 'Sen',
-    color: '#fff',
-    backgroundColor: palette.text.hint,
-    opacity: 0.72,
-    fontSize: '0.75rem',
-    padding: '0 0.5rem',
-    borderRadius: 12,
-  },
-}));
-
-const CustomCard = ({ styles, cover, logo, title, brand, date }) => {
-  const mediaStyles = useCoverCardMediaStyles();
-  return (
-    <Box className={cx(styles.root, styles.color)} pt={20}>
-      <CardMedia classes={mediaStyles} className={styles.cover} image={cover} />
-      <Box className={styles.content} p={2}>
-        <Box position={'relative'} zIndex={1}>
-          <Row gap={2} p={0}>
-            <Item>
-              <Avatar className={styles.logo} src={logo} />
-            </Item>
-            <Item position={'bottom'}>
-              <h2 className={styles.title}>{title}</h2>
-            </Item>
-          </Row>
-          <Row alignItems={'center'} mt={4}>
-            <Item>
-              <div className={styles.team}>{brand}</div>
-            </Item>
-            <Item position={'right'}>
-              <div className={styles.date}>{date}</div>
-            </Item>
-          </Row>
-        </Box>
-      </Box>
-    </Box>
-  );
-};
-
+]
 export const HighlightCardDemo = React.memo(function HighlightCard() {
-  const styles1 = useStyles({ color: '#fc7944' });
-  const styles2 = useStyles({ color: '#5357ce' });
   const gridStyles = useGridStyles();
   return (
     <>
@@ -242,61 +64,21 @@ export const HighlightCardDemo = React.memo(function HighlightCard() {
         style={{ padding: 16 }}
         wrap={'nowrap'}
       >
-        <Grid item>
-          <CustomCard
-            brand={'Overwatch Official'}
-            cover={
-              'https://cdn.vox-cdn.com/thumbor/C6_-SDnnoFdS19XRH4XvAYN1BT8=/148x0:1768x1080/1400x1400/filters:focal(148x0:1768x1080):format(jpeg)/cdn.vox-cdn.com/uploads/chorus_image/image/49641465/tracer_overwatch.0.0.jpg'
-            }
-            date={'02.04.2020'}
-            logo={'https://d3fa68hw0m2vcc.cloudfront.net/bf4/156511609.jpeg'}
-            styles={styles1}
-            title={
-              <>
-                Astronomy Binoculars
-                <br />A Great Alternative
-              </>
-            }
-          />
-        </Grid>
-        <Grid item>
-          <CustomCard
-            brand={'League of Legends Official'}
-            cover={
-              'https://www.pcclean.io/wp-content/uploads/2019/04/559308.jpg'
-            }
-            date={'02.04.2020'}
-            logo={
-              'https://vignette.wikia.nocookie.net/youtube/images/7/77/LeagueOfLegends.jpg/revision/latest?cb=20180718040905'
-            }
-            styles={styles2}
-            title={
-              <>
-                Astronomy Binoculars
-                <br />A Great Alternative
-              </>
-            }
-          />
-        </Grid>
-        <Grid item>
-          <CustomCard
-            brand={'League of Legends Official'}
-            cover={
-              'https://www.pcclean.io/wp-content/uploads/2019/04/559308.jpg'
-            }
-            date={'02.04.2020'}
-            logo={
-              'https://vignette.wikia.nocookie.net/youtube/images/7/77/LeagueOfLegends.jpg/revision/latest?cb=20180718040905'
-            }
-            styles={styles2}
-            title={
-              <>
-                Astronomy Binoculars
-                <br />A Great Alternative
-              </>
-            }
-          />
-        </Grid>
+        {
+          blogs.map((blog, index) => {
+            return  <Grid item key={blog}>
+              <CustomCard
+                description={blog.description}
+                color={blog.color}
+                cover={blog.cover}
+                date={blog.date}
+                userImage={blog.usrImage}
+                title={blog.title}
+              />
+            </Grid>
+          })
+        }
+       
       </Grid>
     </>
   );
