+++ vikmac.pub
@@ -1,0 +1,1 @@
+ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQDH+IO+XPsVKJFhzFz/cXTEf1D5M+z5aXSkH3xGoAeQEHaGIv3yd0O0/QTuD/htPNImp/Rkd8dNPgUht6wV/l1fH5RRcXZS3M3sD8R0LFrhrccSbrycsqpdTHEqZqhNnw1m1qhWTFukoKKIMrKiYcV0YgDrkwfscTb84bUadA52kAeM9nRcK1LU4eut6MdZvaw0hSwsEewfuqYleXdAp40prVGZMlqKgY8kCptFIWn09DlvndV64o36XftSKNOr727f+YH9Le85RtPttExdulUmfqsjTPXMIQxZe6aCe9adr9WwrQYIsAxAYb56BOaZLuUq//1Ats3Wl0U4yNMI9TtBa4S9fgIVpWmZGCHaWNTiOkh8cC23UHAaGXHoAH1oeTxjtCT1ovZjgPyr3B1Jw0hhchrcucdSdNMNYfa+9ICQludi4SdHIC1s5hkJyBnx7Fjj/zkFaHVdhVCK1sx5/cOLiuWWPv9UJqEcRKor3RSbQSkgivVO2gFc+5R/EMC74w0= abizeyimanavictor@victor-abz
